[PROFILING] added debug mode for models and mem plan fixes

+++ test/targets/modules/libs/pulp_nn/include/pulp_nn_kernels.h
@@ -223,7 +223,7 @@ void pulp_nn_pointwise_Co_parallel(
 
 void pulp_nn_linear(
   uint8_t *pInBuffer,
-  int8_t *bias,
+  int *bias,
   uint8_t *pOutBuffer,
   int8_t *pWeights,
   int32_t *k,
